chore: cleaner structure
