+++ apps/cms/src/app/next/seed/collections/data/faqs.ts
@@ -1,0 +1,260 @@
+import { RequiredDataFromCollectionSlug } from "payload";
+
+export const faqs: RequiredDataFromCollectionSlug<"faqs">[] = [
+  {
+    type: "general",
+    question: "How long does it take to receive a response after making an inquiry?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "We will respond to your inquiry within 24 hours.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "general",
+    question: "What payment methods do you accept?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Currently we accept bank transfers.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "general",
+    question: "What are your payment terms?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Generally we accept 30% deposit and 70% balance before shipment.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "general",
+    question: "How long is the lead time?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Generally for made-to-order products, it takes 2-4 weeks for the production and 1-2 weeks for the shipping.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "solar-panel",
+    question: "What is the difference between TOPCon and HJT solar panels?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "TOPCon and HJT are two different types of solar cell technologies. TOPCon uses a tunnel oxide passivated contact, while HJT uses a heterojunction structure. TOPCon is more mainstream and comes at lower cost, but HJT has better bifaciality and higher efficiency.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "solar-panel",
+    question: "How long do solar panels typically last?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Most solar panels have a lifespan of 25 to 30 years, but they can continue to generate electricity at reduced efficiency beyond that period.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "solar-panel",
+    question: "What maintenance do solar panels require?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Solar panels require very little maintenance. It is recommended to keep them clean and free of debris to ensure optimal performance.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+  {
+    type: "solar-panel",
+    question: "What is your MOQ for solar panels?",
+    answer: {
+      root: {
+        type: "root",
+        format: "",
+        indent: 0,
+        version: 1,
+        children: [
+          {
+            type: "paragraph",
+            format: "",
+            indent: 0,
+            version: 1,
+            children: [
+              {
+                mode: "normal",
+                text: "Generally we accept 2 pallets for EU stock products and 1 container for made-to-order products. You can check the details in the product page, or contact us for more information.",
+                type: "text",
+                style: "",
+                detail: 0,
+                format: 0,
+                version: 1,
+              },
+            ],
+          },
+        ],
+        direction: "ltr",
+      },
+    },
+  },
+];
