+++ apps/cms/src/app/next/seed/collections/data/index.ts
@@ -1,0 +1,4 @@
+export * from "./pages";
+export * from "./solar-panels";
+export * from "./faqs";
+export * from "./legal";
